requirements.txt
flask 
web_pdb
pyhthon-dotenv 
mysql-connector-python